// css-diner/src/types/app/app.ts
export enum LvlStatus {
  STATUS_COMPLETED = 'completed',
  STATUS_NOT_COMPLETED = 'not completed',
  STATUS_COMPLETED_WITH_HINT = 'completed with hint',
  STATUS_HINT_USED = 'hint used',
}

export enum StatusToogleRemove {
  TOOGLE = 'toogle',
  REMOVE = 'remove',
}

export enum StatusAddRemove {
  ADD = 'add',
  REMOVE = 'remove',
}

export enum LvlRangeStatus {
  INCREASE = 'increase',
  DECREASE = 'decrease',
}

export enum KeyLocalStorage {
  LOCAL_STORAGE_KEY_LVL = 'lvl',
  LOCAL_STORAGE_KEY_LEVELS_RESULT = 'LevelsResult',
}